backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm>=7"]
build-backend = "setuptools.build_meta"

[project]
name = "arkeon-music-downloader"
version = "3.3.1"
# dynamic = ["readme"]
description = "YouTube Playlist Watcher - Descarga automática a FLAC"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "David Lopez"}
]
keywords = ["youtube", "music", "download", "flac", "playlist", "watcher"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Internet :: WWW/HTTP",
]
requires-python = ">=3.11"
dependencies = [
    "mutagen>=1.47.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "yt-dlp>=2023.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
youtube-watcher = "youtube_watcher.cli:main"

[project.urls]
Homepage = "https://github.com/ArkeonProject/arkeon-music-downloader"
Repository = "https://github.com/ArkeonProject/arkeon-music-downloader"
Issues = "https://github.com/ArkeonProject/arkeon-music-downloader/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.setuptools_scm]
# Deriva versiones desde tags Git (p. ej. v3.0.1)
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"
tag_regex = '^(?:v)?(?P<version>\d+\.\d+\.\d+)$'

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
